pyproject.toml
[project]
name = "gatepass"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "mysqlclient>=2.2.7",
    "psycopg-binary>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
